Comparative Analysis for Games-Related Employment
This document is a comparative analysis of three companies for potential employment in games-related industries: Amazon, Google, and Microsoft. This analysis examines the viability of employment in each of these companies through three criteria, together with a score matrix that I have created, titled Self-Assigned Viability Employment (S.A.V.E) score. Additional findings will be added to this document as further research is conducted and new data emerges.